timeshift: fix compilation error (typecast)

+++ src/timeshift/timeshift_reader.c
@@ -378,7 +378,7 @@ static int _timeshift_do_skip
       cur_file->roff = tsi->pos;
     else
       cur_file->roff = req_time > last_time ? cur_file->size : 0;
-    tvhtrace("timeshift", "do skip cur_file %p roff %"PRId64, cur_file, cur_file->roff);
+    tvhtrace("timeshift", "do skip cur_file %p roff %"PRId64, cur_file, (int64_t)cur_file->roff);
   }
 
   return end;
